requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
crewai==0.28.5
pyyaml==6.0.1
langchain==0.0.335
langchain-openai==0.0.5
openai==1.10.0
